% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_counts.R
\name{add_counts}
\alias{add_counts}
\title{Add counts of multiple columns at once}
\usage{
add_counts(df, ...)
}
\arguments{
\item{df}{the dataframe to work on}

\item{...}{one or multiple column names to summarise by n() on}
}
\value{
a dataframe
}
\description{
Add counts of multiple columns at once
}
\references{
Code originally written in R4DS-Channel by
https://github.com/HannesOberreiter without testing for edge cases,
modified for personal use by me
}
