% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_rowwise.R
\name{apply_rowwise}
\alias{apply_rowwise}
\title{Apply a function rowwise (usually per observation)}
\usage{
apply_rowwise(data, vars, func, na.rm = TRUE)
}
\arguments{
\item{data}{a dataframe}

\item{vars}{a selection of variables from the dataframe to apply a function over}

\item{func}{the function to apply}

\item{na.rm}{TRUE or FALSE, default = TRUE}
}
\value{

}
\description{
Apply a function rowwise (usually per observation)
}
