% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_library.R
\name{clean_library}
\alias{clean_library}
\title{clean library - a function to
set a clean library state - to use with caution}
\usage{
clean_library(lib = .libPaths())
}
\arguments{
\item{lib}{defaults to .libPaths()}
}
\value{
a value
}
\description{
clean library - a function to
set a clean library state - to use with caution
}
