% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Diabetes}
\alias{Diabetes}
\title{Diabetes}
\format{
A sample data frame containing data from 130 probands in 11 variables
}
\usage{
Diabetes
}
\description{
A fake dataframe with variables somewhat related to Diabetes.
Contains character and numeric variables including NAs.
}
\keyword{datasets}
