% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqtable1.R
\name{freqtable1}
\alias{freqtable1}
\title{Gives a frequency table with 1 variable showing absolute'n'relative
frequencies}
\usage{
freqtable1(x, useNA = "n")
}
\arguments{
\item{x}{data and variable to enter (data$variable)}

\item{useNA}{either "always" or "never", abbr. supported, default="n"}
}
\value{
a frequency table with first row absolute numbers, second table
are relative values
}
\description{
Gives a frequency table with 1 variable showing absolute'n'relative
frequencies
}
\examples{
 a <- c(1:5)
 b <- c(11:15)
 sample.data <- data.frame(a,b)
 freqtable1(sample.data$a)

}
