% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matplotlib-colors.R
\docType{data}
\name{matplotlib_colors}
\alias{matplotlib_colors}
\title{Matplotlib categorical color palette}
\format{
An object of class \code{character} of length 10.
}
\usage{
matplotlib_colors
}
\description{
Matplotlib categorical color palette
}
\keyword{datasets}
