% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randucla.R
\name{RAM_char_func}
\alias{RAM_char_func}
\title{summary of all open ended comments - lines with NAs get dropped}
\usage{
RAM_char_func(data, col)
}
\arguments{
\item{data}{a dataframe}

\item{col}{the column with the variables}
}
\value{
a dataframe
}
\description{
summary of all open ended comments - lines with NAs get dropped
}
