% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randucla.R
\name{RAM_table_func}
\alias{RAM_table_func}
\title{iterate the RAM_func over a list}
\usage{
RAM_table_func(topic)
}
\arguments{
\item{topic}{a list to iterate over}
}
\value{
a dataframe
}
\description{
iterate the RAM_func over a list
}
