% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableone_mat_to_data_frame.R
\name{tableone_mat_to_data_frame}
\alias{tableone_mat_to_data_frame}
\title{Convert a tableone matrix to a dataframe}
\usage{
tableone_mat_to_data_frame(mat)
}
\arguments{
\item{mat}{An object created through tableone()}
}
\value{
a dataframe
}
\description{
Convert a tableone matrix to a dataframe
}
\references{
developed & published in a gist from Kazuki Yoshida
https://gist.github.com/kaz-yos/7b4e97b633d61fa4d1d27f46f2534ae3
}
