% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tableone_mat_to_excel.R
\name{write_tableone_mat_to_xlsx}
\alias{write_tableone_mat_to_xlsx}
\title{Write tableone mat to excel file}
\usage{
write_tableone_mat_to_xlsx(tableone_mat, file)
}
\arguments{
\item{tableone_mat}{a tableone-object}

\item{file}{the name of the file.xlsx}
}
\value{

}
\description{
Write tableone mat to excel file
}
\references{
developed & published in a gist from Kazuki Yoshida
https://gist.github.com/kaz-yos/7b4e97b633d61fa4d1d27f46f2534ae3
}
